Add test case for replacing a directory with a file
Consider a target directory containing a file `dir/file` and a source
directory containing a file `dir`. Note that in the target `dir` is a directory
and in the source `dir` is file. Currently `kudusync` does not remove the
directory in the target folder before attempting to create the file `dir`.

This PR adds a test case to show the problem. It does not fix the problem.